+++ templates/nvim/lua/workspace/core/options.lua
@@ -29,8 +29,32 @@ opt.signcolumn = "yes" -- show sign column so that text doesn't shift
 -- backspace
 opt.backspace = "indent,eol,start" -- allow backspace on indent, end of line or insert modce start position
 
--- clipboard
+-- clipboard - use OSC 52 for container/SSH clipboard support
 opt.clipboard:append("unnamedplus") -- use system clipboard as default register
+
+-- OSC 52 clipboard provider for containers
+-- This allows yanking to host clipboard over terminal escape sequences
+-- Works with: WezTerm, iTerm2, Kitty, Alacritty, Windows Terminal
+if os.getenv("CONTAINER") or os.getenv("SSH_TTY") or os.getenv("DVM_WORKSPACE") then
+  local function paste()
+    return {
+      vim.fn.split(vim.fn.getreg(""), "\n"),
+      vim.fn.getregtype(""),
+    }
+  end
+
+  vim.g.clipboard = {
+    name = "OSC 52",
+    copy = {
+      ["+"] = require("vim.ui.clipboard.osc52").copy("+"),
+      ["*"] = require("vim.ui.clipboard.osc52").copy("*"),
+    },
+    paste = {
+      ["+"] = paste,
+      ["*"] = paste,
+    },
+  }
+end
 
 -- split windows
 opt.splitright = true -- split vertical window to the right